Transformations_Backend-main/requirements.txt
blinker==1.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
greenlet==3.0.3
httplib2==0.20.2
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
polars==0.20.15
psycopg2-binary==2.9.9
pyarrow==15.0.1
pycairo==1.26.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.28
systemd-python==234
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
zipp==1.0.0
